URL, edgeURL, and labelURL support
`edgeURL` and `labelURL` are technically already supported via the `**attrs` Dict in `Edge`'s initializer.

I added a named parameter to `Node` to accept [`URL`](https://www.graphviz.org/doc/info/attrs.html#a:URL).

SVG output natively supports links in the diagram. Links can also be added to jpg and png using `cmapx` which I added to the output types.

To use cmapx you will need to run the diagram generation twice. Once for the png and once for the cmapx file.

I had originally detected if a url was added to a node, edge, or label and autogenerated the cmapx file. I can add that back in if interested, but it seemed to break the paradigm of generating _one file_ from a diagram.